Step 7: We then create our vue component, doing what ever we need. This example imports an free "password strength" package which we interface with.
We are using a "watch" on line 25 to update the actual form element with the value returned from our Vue component.

We should now have a fully working custom field built with vue. If you're still running `npm run watch` this should have loaded for you.

We can confirm it's working in the next step.